requirements.txt
streamlit
tensorflow
opencv-python
Pillow
numpy
scikit-learn
